Add output template which restores the BBCode
Currently, this template tries to restore the original BBCode: The result is not 100% equal to the original,as the parser adds closing tags (if they are optional) and might modify tags (e.g. the media tag is converted into the actual tag from the given URL)

Later this template will be used to provide error feedback on the provided BBCode.