Use iframe-API for YouTube videos (closes #18)
The newer API (https://developers.google.com/youtube/iframe_api_reference) is more flexible and supports a wider range of devices.

Note that iframe-tag is directly inserted, instead of using the JavaScript code.
This is easier to generate and include on pages that support BBCode, and might or might now use youtube videos